Refactoring: wait_for() moved from JobQueue to jobs.await_job()
Rationale: The current async JobQueue implementation just polls the repository, which might as well be the default queue-independent implementation. In theory, a JobQueue might have a more efficient way of awaiting a job, but the AsyncJobQueue hasn't (since a job could run as coroutine in an entirely different worker process) and moving that code into a service function simplifies the job interface.